Fix incorrect preposition usage.
The ltyp field was dropped in portal_branches (and thus in 0.11+), and this
meant that Sequell was always using "in".

As far as I can tell, the ltyp field was only being used to make it say
"in Zig:14" rather than "on Zig:14", and the latter seems arguably more
correct anyway.