Throttle the http service sanely.
Instead of naive locking, keep track of the number of active requests and
immediately refuse overloading requests with a 503, since our clients are smart
enough to retry.

This fixes the pathological failure mode where the service wastes time
processing requests that were issued in the distant past where the client has
stopped waiting for an answer.