Add support for end-of-line comments to !lg
It'd be nice to be able to write like `!lm * recent won nrune=3 lg:nrune=3 s=rune -- which rune is usually taken third in a 3-rune game` to explain the context of a query. Currently if I want to run a query and tell someone (or the room) about it, the only way to explain it is to send a separate irc message, which is a little tacky.

I don't actually have an instance of sequell to test this on, and gave up on installing all of the necessary dependencies, but I think the change is simple enough that it ought to work.