Add class mapping for Hedge Wizard
The old Wizard class become Hedge Wizard, and we've considered them as separate classes because of differences between starting equipment and stats. However we've not yet updated the class abbreviation, so for now Wz = Hedge Wizard. Update the Sequell data accordingly. If we end up renaming the Hedge Wizard abbreviation to something like HW, we may need to do a retroactive combo rename in Sequell. If we instead keep the Wz abbreviation for Hedge Wizard, we may instead need to retroactively rename Wizard combos to use an abbreviation like Wi.