feat: Retroactively rename Wizard to Hedge Wizard
Based on the entries in this file made for the Assassin->Brigand retroactive
rename, these changes should be sufficient to rename Wizard to Hedge Wizard in
terms of the Sequell configuration. greensnark will need to do updates on the
database side of things and can hopefully confirm that I've made the right
changes here.