task: Add On and Al, and mark Og and VM as retired
The previous attempt at adding these failed because `On` is interpreted
as `true` in yaml 1.1 and earlier. It seems we can't upgrade all of
Sequell's dependencies to support yaml 1.2, where these alternative
interpretations are removed. Namely because Perl's YAML::XS doesn't seem
to have yaml 1.2 support. So this commit simply encloses the On key in
quotes.

This reverts commit 61b73b081ae056b8547d6c7793d368f9dc970f10.