feat: Remove the 0.31 tournament file filter
A special filter on the source file is often needed for CWZ, which may
not have trunk updated to the next alpha version. In this case, all CWZ
games during the tournament period will be 0.32-a, since CWZ updated
trunk before tournament start (and it always force-transfers trunk
games). Sadly CWZ itself still isn't in the tournament, but if it does
join, at least it will be properly 0.31.0. In any case, we can remove
the file filter to speed up queries that use the `t` keyword.