task: Clean up sources
This commit marks various retired sources as such by removing the * at
the end of their entries. This will prevent Sequell from attempting to
fetch these as part of its usual cycle; it will only fetch them if it
didn't have a local copy of the log.

Additionally we order the sources so that all servers with active
entries have those entries at the top of the listing. We also rework
some source lines to have a new (retired) entry trunk zotdef, since that
game type no longer works. A few other sources are reordered, cleaned
up, and combined as necessary.