task: Some crawl and source cleanups
This commit is mostly to try to poke Sequell into reloading its sources
so that it can see the 0.32 games. The tournament is close to starting,
but it's still not seeing those sources. The changes are to add Crucible
to the branch list, add an alias for Alchemy skill, and disable updates
to CWZ, now that it's offline.